tasks/mnist_classifier/FileNest/tasks/mnist_classifier/MNIST_Classifier/FileNest/tasks/mnist_classifier/MNIST_Classifier/pyproject.toml
[project]
name = "mnist_classifier"
version = "0.1.0"
dependencies = [
    "torch",
    "torchvision",
    "wandb",
    "hydra-core",
    "ray[tune]",
    "torchinfo",
    "black",
    "isort"
]
